refactor: exceptions e services
